[MOD] Restored WS queries, once more...
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@21647 b456876b-0849-0410-b77d-98878d47e9d5

+++ db/tiki-4.0-pgsql.sql
@@ -471,7 +471,7 @@ CREATE TABLE "tiki_categories" (
   "description" varchar(250) default NULL,
   "parentId" bigint default NULL,
   "hits" integer default NULL,
-  "rootCategId" integer default NULL,
+  "rootCategId" bigint default NULL,
   PRIMARY KEY ("categId")
 ) ;
