+++ db/tiki-4.0-sqlite.sql
@@ -475,7 +475,7 @@ CREATE TABLE tiki_categories (
   description varchar(250) default NULL,
   parentId bigint default NULL,
   hits integer default NULL,
-  rootCategId integer default NULL,
+  rootCategId bigint default NULL,
   PRIMARY KEY (categId)
 ) ENGINE=MyISAM ;
